Changed json-schema Validation type to 'anyOf'
Validation type of 'oneOf' is not correct for our usage, as it requires exactly one match out of the array.  In our translation from CSDL, these are either full schema inclusions or just link references, but in all cases, both of these will be valid.  And in the future, with additional versioned schema files in this list, it is very likely a payload could validate to multiple versions of the same schema.