Add OPTIONS method to prevent CORs from being forbidden ( Issue #989 ).
The lack of support for the OPTIONS method hinders the CORs capabilities of Redfish. This is all related to the CORs preflight protocol outlined in the W3C CORs spec. The only CORs mentions presently are the Origin and Access-Control-Allow-Origin headers. These are enough to grant a web-app read-only access to Redfish via CORs so long as no authorization is required. Anything beyond that would require CORs preflight.